move _operator classes to operator
This matches what they call themselves at runtime
